Fix compiler warnings about sign conversion and type mismatches
- src/radix_tree.c: Cast bitwise operations to uint16_t to avoid sign conversion warnings
- src/landlock_bridge.c: Cast array size calculation to size_t
- src/rule_engine.c: Fix format string truncation warnings and cast hash values to uint32_t
- tests/mock_fs.c: Cast memcpy length to size_t
- tests/test_policy_parser.c: Cast byte values to char in XOR operations
- tests/test_radix_tree_edge.c: Change pos to size_t and handle snprintf return value properly
- tests/test_radix_tree_extended.c: Change pos to size_t, use unsigned int for loop, and handle snprintf return value
- tests/test_compilation_binary.c: Cast unused return values to void to silence warnings

+++ rbox-ruleset/src/landlock_bridge.c
@@ -427,7 +427,7 @@ landlock_builder_t *soft_ruleset_to_landlock(const soft_ruleset_t *rs,
                     if (deny_idx >= deny_cap) {
                         deny_cap = deny_cap == 0 ? 16 : deny_cap * 2;
                         deny_prefixes = realloc(deny_prefixes,
-                                                (deny_cap + 1) * sizeof(char *));
+                                                (size_t)(deny_cap + 1) * sizeof(char *));
                         if (!deny_prefixes) {
                             landlock_builder_free(b);
                             soft_landlock_deny_prefixes_free(deny_prefixes);
@@ -633,7 +633,7 @@ landlock_builder_t *soft_ruleset_to_landlock_with_report(
                     if (deny_idx >= deny_cap) {
                         deny_cap = deny_cap == 0 ? 16 : deny_cap * 2;
                         deny_prefixes = realloc(deny_prefixes,
-                                                (deny_cap + 1) * sizeof(char *));
+                                                (size_t)(deny_cap + 1) * sizeof(char *));
                         if (!deny_prefixes) {
                             landlock_builder_free(b);
                             soft_landlock_deny_prefixes_free(deny_prefixes);
